Code refactor with useWPBodyMarginOffsetEffect.
